requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.4
packaging==23.2
requests==2.31.0
urllib3==2.1.0
paho-mqtt==2.0.0